--- src/components/Header/123.jsx
@@ -1,51 +1,0 @@
-// import { FaSearch } from "react-icons/fa";
-// import styled from "styled-components";
-
-// const SearchForm = () => {
-//   return (
-//     <StForm>
-//       <input type="text" />
-//       <StIcon />
-//     </StForm>
-//   );
-// };
-
-// export default SearchForm;
-
-// const StForm = styled.form`
-//   display: flex;
-//   align-items: center;
-//   position: absolute;
-//   left: 50%;
-//   transform: translateX(-50%);
-//   width: 35%;
-
-//   input {
-//     border: none;
-//     width: 100%;
-//     height: 33px;
-//     border-radius: 6px;
-//     background-color: var(--color-black-10);
-//     outline: none;
-//     padding: 0 10px;
-//   }
-//   @media (max-width: 860px) {
-//     & {
-//       position: static;
-//       transform: none;
-//     }
-//   }
-//   @media (max-width: 700px) {
-//     & {
-//       width: 230px;
-//       /* display: none; */
-//     }
-//     @media (max-width: 600px) {
-//       display: none;
-//     }
-//   }
-// `;
-// const StIcon = styled(FaSearch)`
-//   margin-left: 5px;
-//   font-size: 2rem;
-// `;
